Create xchat_rename_ugly_contacts.py
This is a script for Xchat that is very useful if you use bitlbee and facebook chat or google chat since facebook contacts would have the nickname like this for example "-100000489402714" (leading dash then facebook profile number). So this script renames the users when they join the chat the first time.